Moved logic to a hook, small update and added moreinformation component
